Separate "campaign parameters" and "thankyou campaign parameters"
- the usual banner campaign parameters should not have to care about thankyou campaign related things

https://phabricator.wikimedia.org/T384693
