+++ banners/thank_you/banner_var.wpde.ts
@@ -40,7 +40,7 @@ const app = createVueApp( BannerConductor, {
 	},
 	bannerCategory: 'fundraisingThankyou',
 	bannerProps: {
-		settings: createThankYouSettings( new IntegerDe(), page.getCampaignParameters().thankYouCampaign ),
+		settings: createThankYouSettings( new IntegerDe(), page.getThankYouCampaignParameters() ),
 		subscribeURL: createTrackedURL( SUBSCRIBE_URL, page.getTracking(), impressionCount, Locales.DE ),
 		useOfFundsURL: createTrackedURL( USE_OF_FUNDS_URL, page.getTracking(), impressionCount, Locales.DE )
 	},
