+++ dashboard/wpde-offline/index.html
@@ -80,11 +80,13 @@
 		endDate: '2024-12-31',
 		numberOfMembers: 73832,
 		isLateProgress: true,
-		dramaTextIsVisible: false,
-		thankYouCampaign: {
-			numberOfDonors: 350000,
-			progressBarPercentage: 100
-		}
+		dramaTextIsVisible: false
+	};
+
+	window.thankYouCampaignParameters = {
+		numberOfDonors: 350000,
+		numberOfMembers: 73832,
+		progressBarPercentage: 100
 	};
 
 	// drop-in replacement for dev-mode-wpde.js "banner" from bruce, without jQuery
