+++ src/domain/CampaignParameters.ts
@@ -14,12 +14,6 @@ export interface CampaignProjectionParameters {
 	averageAmountPerDonation: number
 }
 
-export interface ThankYouCampaignParameters {
-	numberOfDonors: number;
-	numberOfMembers: number;
-	progressBarPercentage: number;
-}
-
 /**
  * Campaign parameters is a value object with string and number primitives used to pass values from
  * the "environment" of the banner (i.e. wikipedia.org or wikipedia.de) to the dynamic text rendering.
@@ -41,6 +35,4 @@ export interface CampaignParameters {
 	isLateProgress: boolean,
 	dramaTextIsVisible: boolean,
 	urgencyMessageDaysLeft: number,
-
-	thankYouCampaign: ThankYouCampaignParameters
 }
