+++ src/domain/ThankYouCampaignParameters.ts
@@ -1,0 +1,5 @@
+export interface ThankYouCampaignParameters {
+	numberOfDonors: number;
+	numberOfMembers: number;
+	progressBarPercentage: number;
+}
