+++ test/unit/utils/DynamicContent/DynamicCampaignText.spec.ts
@@ -41,12 +41,7 @@ const campaignParameters: CampaignParameters = {
 	startDate: '2023-11-03',
 	urgencyMessageDaysLeft: 10,
 	isLateProgress: false,
-	dramaTextIsVisible: false,
-	thankYouCampaign: {
-		progressBarPercentage: 80,
-		numberOfDonors: 42,
-		numberOfMembers: 23
-	}
+	dramaTextIsVisible: false
 };
 const impressionCount: ImpressionCount = {
 	bannerCount: 42,
